+++ .gitignore
@@ -40,7 +40,3 @@ build/
 
 ### VS Code ###
 .vscode/
-<<<<<<< HEAD
-
-=======
->>>>>>> origin
