+++ pom.xml
@@ -92,8 +92,13 @@
 			<version>4.12</version><!--$NO-MVN-MAN-VER$-->
 			<scope>test</scope>
 		</dependency>
+        <dependency>
+            <groupId>com.konghq</groupId>
+            <artifactId>unirest-java</artifactId>
+            <version>3.7.04</version>
+        </dependency>
 
-	</dependencies>
+    </dependencies>
 
 	<build>
 		<plugins>
