+++ src/main/java/com/revature/backend/controller/SwotController.java
@@ -1,5 +1,10 @@
 package com.revature.backend.controller;
 
+import static com.revature.backend.util.ClientMessageUtil.CREATION_FAILED;
+import static com.revature.backend.util.ClientMessageUtil.DELETION_FAILED;
+import static com.revature.backend.util.ClientMessageUtil.SUCCESSFULLY_CREATED;
+import static com.revature.backend.util.ClientMessageUtil.SUCCESSFULLY_DELETED;
+
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,20 +13,17 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.backend.model.AnalysisItem;
 import com.revature.backend.model.Swot;
 import com.revature.backend.service.SwotService;
 import com.revature.backend.util.ClientMessage;
-import static com.revature.backend.util.ClientMessageUtil.*;
-
-//TODO fix imports for optimization
 
 @RestController("swotController")
 @RequestMapping("/swot")
@@ -30,50 +32,110 @@ public class SwotController {
 	@Autowired
 	private SwotService swotService;
 	
-	//create -swot
+	/*
+	 * POST request to create a new SWOT for an Associate
+	 * with AnalysisItems within it as a collection.
+	 * 
+	 * Takes in a Swot as a nested JSON object: the AnalysisItems
+	 * will be found within a JSON array.
+	 * 
+	 * Returns a 201 status with a message for the client
+	 * within the HTTP body as a JSON object if successful.
+	 * 
+	 * *COMPLETED*
+	 */
 	@PostMapping(path="/create", consumes= {MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<ClientMessage> createSwot(@RequestBody Swot swot) {
 		ClientMessage body = swotService.createNewSwot(swot) ? SUCCESSFULLY_CREATED : CREATION_FAILED;
-//		return ResponseEntity.ok(body); //TODO this should be 201: CREATED
-		return ResponseEntity.status(HttpStatus.CREATED).body(body); //needs testing
+		return ResponseEntity.status(HttpStatus.CREATED).body(body);
 	}
 	
-	//get all -swot
-	@GetMapping(path="/view", consumes= {MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<List<Swot>> viewSwot(@RequestParam("associateId") int associateId) {
-		List<Swot> swotList = swotService.retrieveAllSwotByAssociateID(associateId);
+	/*
+	 * GET request for fetching all SWOTs based on an 
+	 * Associate's id as found in the RESTful URL.
+	 * 
+	 * Returns the List of SWOTs as a JSON array with
+	 * a 200 code if successful.
+	 * 
+	 * *COMPLETED*
+	 */
+	@GetMapping(path="/view/{associateId}", produces= {MediaType.APPLICATION_JSON_VALUE})
+	public ResponseEntity<List<Swot>> viewSwot(@PathVariable("associateId") int associateId) {
+		List<Swot> swotList = swotService.retrieveAllSwotByAssociateId(associateId);
 		return ResponseEntity.ok(swotList);
 	}
 	
-	//get all -swot
+	/*
+	 * GET request for fetching all existing SWOTs in
+	 * the system. Returns the List as a JSON array.
+	 * 
+	 * This will more than likely be removed in the
+	 * future and was only used for proof of concept.
+	 * 
+	 *  *COMPLETED BUT FIT FOR REMOVAL*
+	 */
 	@GetMapping(path="/view/all", produces= {MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<List<Swot>> viewAllSwot() {
 		List<Swot> swotList = swotService.retrieveAllSwot();
 		return ResponseEntity.ok(swotList);
 	}
 	
-	//create -item
+	/*
+	 * POST request to create a new AnalysisItem for a
+	 * particular SWOT. 
+	 * 
+	 * Takes in an AnalysisItem from a
+	 * JSON object. 
+	 * 
+	 * Returns a 201 HTTP status with an 
+	 * informative client message in the body.
+	 * 
+	 * *COMPLETED*
+	 */
 	@PostMapping(path="/item/new", consumes= {MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<ClientMessage> createItem(@RequestBody AnalysisItem analysisItem) {
 		ClientMessage body = swotService.createNewItem(analysisItem) ? SUCCESSFULLY_CREATED : CREATION_FAILED;
-		return ResponseEntity.ok(body); //TODO this should be 201: CREATED
+		return ResponseEntity.status(HttpStatus.CREATED).body(body);
 	}
 	
-	//update -item
-	@PutMapping(path="/item/update", consumes= {MediaType.APPLICATION_JSON_VALUE})
+	/*
+	 * PUT request to update an AnalysisItem.
+	 * 
+	 * Takes in an AnalysisItem *with* altered fields
+	 * from a JSON object.
+	 * 
+	 * If the update failed, returns a 204 status code.
+	 * 
+	 * If successful, returns a 200 status code and the
+	 * updated object.
+	 * 
+	 * *COMPLETED*
+	 */
+	@PutMapping(path="/item/update/{analysisItemId}", consumes= {MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<AnalysisItem> updateItem(@RequestBody AnalysisItem analysisItem) {
 		AnalysisItem updatedItem = swotService.updateItem(analysisItem);
-		if (updatedItem == null) {
-			return ResponseEntity.badRequest().body(null);	// Update likely failed due to misinput of information.
-		}else {
+		if (updatedItem == null) {	// A null object denotes that the update had failed or done nothing.
+			return ResponseEntity.noContent().build();	// 204 would be the proper response for a PUT that changes nothing.
+		}else{	// Success, return the item.
 			return ResponseEntity.ok(updatedItem);
 		}
 	}
 	
-	//delete -item
-	@DeleteMapping(path="/item/delete", consumes= {MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<ClientMessage> createSwot(@RequestBody AnalysisItem analysisItem) {
-		ClientMessage body = swotService.deleteItem(analysisItem) ? SUCCESSFULLY_DELETED : DELETION_FAILED;
+	/*
+	 * DELETE request to delete an AnalysisItem from
+	 * a particular SWOT.
+	 * 
+	 * Takes in the id of the AnalysisItem as a
+	 * variable from the URL.
+	 * 
+	 * Returns an informative client message and
+	 * a 200 status code.
+	 * 
+	 * *COMPLETED*
+	 */
+	@DeleteMapping(path="/item/delete/{analysisItemId}")
+	public ResponseEntity<ClientMessage> createSwot(@PathVariable("analysisItemId") int analysisItemId) {
+		ClientMessage body = swotService.deleteItem(analysisItemId) ? SUCCESSFULLY_DELETED : DELETION_FAILED;
 		return ResponseEntity.ok(body);
 	}
 }
