+++ src/main/java/com/revature/backend/repository/SwotRepository.java
@@ -11,7 +11,8 @@ import com.revature.backend.model.Swot;
 @Repository
 public interface SwotRepository extends JpaRepository<Swot, Integer> {
 
-	//DBG if this crashes use associate_id instead of associate TODO: remove this
-	@Query(value = "select * from swot_analysis where associate =?1", nativeQuery = true)
-	List<Swot> findAllByAssociateID(int associateId);
+	// TODO: There was a conflict here and I wasn't sure which one was correct.
+//	@Query(value = "select * from swot_analysis where associate =?1", nativeQuery = true)
+	@Query(value = "select * from swot_analysis where associate_id =?1", nativeQuery = true)
+	List<Swot> findAllByAssociateId(int associateId);
 }
