+++ src/main/resources/application.properties
@@ -1,9 +1,20 @@
 logging.level.org.springframework.data=INFO
 logging.level.org.springframework.jdbc.core.JdbcTemplate=DEBUG
 
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-spring.h2.console.enabled=true
+#spring.datasource.url=jdbc:h2:mem:testdb
+#spring.datasource.driverClassName=org.h2.Driver
+#spring.datasource.username=sa
+#spring.datasource.password=password
+#spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+#spring.h2.console.enabled=true
+
+spring.datasource.hikari.connectionTimeout=20000
+spring.datasource.hikari.maximumPoolSize=5
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+#spring.datasource.url=jdbc:postgresql://datarevdatabase.ceqrfdqfbwyy.us-east-2.rds.amazonaws.com/pokemon
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+spring.jpa.hibernate.ddl-auto=update
