+++ src/main/resources/testData.sql
@@ -1,0 +1,4 @@
+INSERT INTO MANAGER VALUES  (1,'Testemail@email.email','test','user');
+INSERT INTO MANAGER VALUES  (2,'mail@mail.email','testy','usery');
+INSERT INTO MANAGER VALUES  (3,'test@email.mail','isa','cat');
+
