+++ src/test/java/com/revature/backend/testing/service/StagingListenerTesting.java
@@ -14,32 +14,53 @@ import com.revature.backend.model.api.ApiBatchTemplate;
 import com.revature.backend.util.StagingListener;
 import com.revature.backend.util.StagingListenerImpl;
 
-@SpringBootTest(classes={StagingListener.class, StagingListenerImpl.class})
+@SpringBootTest(classes = { StagingListener.class, StagingListenerImpl.class })
 @RunWith(SpringRunner.class)
 public class StagingListenerTesting {
 
 	@Autowired
 	StagingListenerImpl stagingListener;
-	
-	
+
 	@Test
-	void testGetNewBatches()
-	{
+	void testGetNewBatches() {
 		stagingListener.checkForNewBatches();
 		List<ApiBatchTemplate> batches = stagingListener.getLatestBatches();
 		for (ApiBatchTemplate apiBatchTemplate : batches) {
 			System.out.println(apiBatchTemplate.toString());
 		}
-		if(stagingListener.triggerUpdate())
-		{
-			assertTrue(batches.size() >0);
-		}
-		else {
+		if (stagingListener.triggerUpdate()) {
+			assertTrue(batches.size() > 0);
+		} else {
 			assertTrue(batches.size() == 0);
 		}
 	}
-	
-	
-	
-	
+
+	@Test
+	void testGetLatestBatches() {
+		stagingListener.checkForNewBatches();
+		List<ApiBatchTemplate> batches = stagingListener.getLatestBatches();
+		if (stagingListener.triggerUpdate()) {
+			assertTrue(batches.size() > 0);
+		} else {
+			assertTrue(batches.size() == 0);
+		}
+	}
+
+	@Test
+	void testGetLatestBatchesWithFalseBatchesInserted() {
+		stagingListener.mockCheckForNewBatches(true);
+		List<ApiBatchTemplate> batches = stagingListener.getLatestBatches();
+
+		assertTrue(batches.size() > 0);
+
+	}
+
+	@Test
+	void testGetLatestBatchesWithNoBatchesInserted() {
+		stagingListener.mockCheckForNewBatches(false);
+		List<ApiBatchTemplate> batches = stagingListener.getLatestBatches();
+		assertTrue(batches.size() == 0);
+
+	}
+
 }
